Add files via upload
Sorting is a process of ordering individual elements of a list according to their proper rank, either in ascending or descending order. We can easily sort a list of elements by means of iteration /loop and if-condition check statements. Sorting algorithms can be implemented by any programming languages
Bubble Sort- A sorting algorithm which compares one element to its next element and if requires it swaps like a bubble.
Selection Sort - A sorting algorithm which selects a position in the elements and compares it to the rest of the positions one by one.
Insertion Sort - a sorting algorithm that places an unsorted element at its suitable place in each iteration.